refactor(routes): move register routes into its own router
