+++ src/routers/registerRouter.ts
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import {
+	registerUserGet,
+	registerUserPost,
+} from "../controllers/registerController.js";
+
+const registerRouter = Router();
+
+registerRouter.get("/", registerUserGet);
+registerRouter.post("/", ...registerUserPost);
+
+export default registerRouter;
